ai_stock_business/requirements.txt
replicate>=1.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
paramiko>=3.3.0
Pillow>=10.0.0